Break Sync Engine into multiple files
